+++ pkg/annotation/machinedeployment.go
@@ -1,0 +1,5 @@
+package annotation
+
+// MachineDeploymentSubnet is used to specify the IPv4 CIDR used for a
+// node pool defined by a MachineDeployment or AWSMachineDeployment CR.
+const MachineDeploymentSubnet = "machine-deployment.giantswarm.io/subnet"